fix: support half-hour and quarter-hour time zones
Timezone was an int and the offset was computed as 3600 * Timezone, so
zones like India (+5:30), Nepal (+5:45), Iran (+3:30) and central
Australia (+9:30) could not be expressed.

Makes the setting fractional and rounds to the nearest second when
applying it. Existing configs keep working: ArduinoJson converts a stored
integer to float, and %g prints whole zones without a decimal part. The
config portal field grows from 3 to 6 characters so "-9.5" fits.

Reported in BitMaker-hub/NerdMiner_v2#738.

+++ src/drivers/storage/nvMemory.cpp
@@ -96,7 +96,7 @@ bool nvMemory::loadConfig(TSettings* Settings)
                     if (json.containsKey(JSON_SPIFFS_KEY_POOLPORT))
                         Settings->PoolPort = json[JSON_SPIFFS_KEY_POOLPORT].as<int>();
                     if (json.containsKey(JSON_SPIFFS_KEY_TIMEZONE))
-                        Settings->Timezone = json[JSON_SPIFFS_KEY_TIMEZONE].as<int>();
+                        Settings->Timezone = json[JSON_SPIFFS_KEY_TIMEZONE].as<float>();
                     if (json.containsKey(JSON_SPIFFS_KEY_STATS2NV))
                         Settings->saveStats = json[JSON_SPIFFS_KEY_STATS2NV].as<bool>();
                     if (json.containsKey(JSON_SPIFFS_KEY_INVCOLOR)) {
